+++ data/src/main/kotlin/data/notification/NotificationManager.kt
@@ -9,26 +9,26 @@ internal interface NotificationManager {
                @StringRes title: Int,
                @StringRes body: Int,
                @NotificationCategory category: String,
-               @NotificationPriority priority: Long = NotificationManager.PRIORITY_MEDIUM,
-               @NotificationVisibility visibility: Long = NotificationManager.VISIBILITY_PUBLIC,
+               @NotificationPriority priority: Int = NotificationManager.PRIORITY_MEDIUM,
+               @NotificationVisibility visibility: Int = NotificationManager.VISIBILITY_PUBLIC,
                clickHandler: PendingIntent? = null)
 
     fun notify(@StringRes channelName: Int,
                title: String,
                body: String,
                @NotificationCategory category: String,
-               @NotificationPriority priority: Long = NotificationManager.PRIORITY_MEDIUM,
-               @NotificationVisibility visibility: Long = NotificationManager.VISIBILITY_PUBLIC,
+               @NotificationPriority priority: Int = NotificationManager.PRIORITY_MEDIUM,
+               @NotificationVisibility visibility: Int = NotificationManager.VISIBILITY_PUBLIC,
                clickHandler: PendingIntent? = null)
 
     companion object {
         const val CATEGORY_RECOMMENDATION = NotificationCompat.CATEGORY_RECOMMENDATION
         const val CATEGORY_SERVICE = NotificationCompat.CATEGORY_SERVICE
-        const val PRIORITY_MEDIUM = NotificationCompat.PRIORITY_DEFAULT.toLong()
-        const val PRIORITY_LOW = NotificationCompat.PRIORITY_LOW.toLong()
-        const val PRIORITY_HIGH = NotificationCompat.PRIORITY_HIGH.toLong()
-        const val VISIBILITY_PRIVATE = NotificationCompat.VISIBILITY_PRIVATE.toLong()
-        const val VISIBILITY_PUBLIC = NotificationCompat.VISIBILITY_PUBLIC.toLong()
-        const val VISIBILITY_SECRET = NotificationCompat.VISIBILITY_SECRET.toLong()
+        const val PRIORITY_MEDIUM = NotificationCompat.PRIORITY_DEFAULT
+        const val PRIORITY_LOW = NotificationCompat.PRIORITY_LOW
+        const val PRIORITY_HIGH = NotificationCompat.PRIORITY_HIGH
+        const val VISIBILITY_PRIVATE = NotificationCompat.VISIBILITY_PRIVATE
+        const val VISIBILITY_PUBLIC = NotificationCompat.VISIBILITY_PUBLIC
+        const val VISIBILITY_SECRET = NotificationCompat.VISIBILITY_SECRET
     }
 }
